Use proxyarp app instead of multicast groups
We did not manage to handle ARPs entirely in the data plane (without
nasty workarounds), as FPM does not support a pipeline behavior where
for the same packet we set a multicast group ID and we clone to the CPU,
since both actions are currently mapped to the same table.

We drop using multicast replication in the P4 program for ARPs and
instead rely on the proxyarp ONOS app to handle ARPs via packet-in/out.